Delete chatboxviewmodelfactory
